+++ examples/eventmanager/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/ormushq/ormus/contract/go/project"
 	"github.com/ormushq/ormus/eventmanager"
 	"github.com/ormushq/ormus/pkg/channel"
-	rbbitmqchannel "github.com/ormushq/ormus/pkg/channel/adapter/rabbitmq"
+	"github.com/ormushq/ormus/pkg/channel/adapter/rabbitmqchannel"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
@@ -22,7 +22,7 @@ func main() {
 	port := 5672
 	reconnectSecond := 2
 	// Init channel package you can use rabbitmqchannel or simple
-	rc := rbbitmqchannel.New(done, wg, rbbitmqchannel.Config{
+	rc := rabbitmqchannel.New(done, wg, rabbitmqchannel.Config{
 		User:            "guest",
 		Password:        "guest",
 		Host:            "127.0.0.1",
