+++ manager/delivery/httpserver/userhandler/register_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/ormushq/ormus/manager/workers"
 	"github.com/ormushq/ormus/param"
 	"github.com/ormushq/ormus/pkg/channel"
-	"github.com/ormushq/ormus/pkg/channel/adapter/simple"
+	"github.com/ormushq/ormus/pkg/channel/adapter/simplechannel"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -66,7 +66,7 @@ func TestIntegrationHandler_Register(t *testing.T) {
 	cfg := config.C().Manager
 	done := make(chan bool)
 	wg := sync.WaitGroup{}
-	internalBroker := simple.New(done, &wg)
+	internalBroker := simplechannel.New(done, &wg)
 	internalBroker.NewChannel("CreateDefaultProject", channel.BothMode,
 		cfg.InternalBrokerConfig.ChannelSize, cfg.InternalBrokerConfig.MaxRetryPolicy)
 	repo := usermock.NewMockRepository(false)
