+++ pkg/channel/adapter/simplechannel/adapter.go
@@ -1,4 +1,4 @@
-package simple
+package simplechannel
 
 import (
 	"fmt"
