+++ pkg/channel/adapter/simplechannel/channel.go
@@ -1,4 +1,4 @@
-package simple
+package simplechannel
 
 import (
 	"sync"
